requirements.txt
streamlit
python-dotenv
langchain
langchain_community
PyPDF2
faiss-cpu
langchain-openai
langserve
fastapi
uvicorn
sse_starlette


